move act to its own module
