--- act/Examples/AmmGenerated.hs
@@ -1,13 +1,0 @@
-{-# LANGUAGE DataKinds #-}
-{-# LANGUAGE QuasiQuotes #-}
-{-# LANGUAGE TemplateHaskell #-}
-{-# LANGUAGE TypeApplications #-}
-
-module Examples.AmmGenerated where
-
-import Act.TH
-
--- This generates the `ammContract`
-$(act2OG "amm.act")
-
--- $(hevm2OG "contract") --- do something???
