+++ act/README.md
@@ -1,0 +1,4 @@
+# Act integration of open games
+
+This project implements the integration of open games with the ACT project. The goal is to be able to
+perform game-theoretic analysis of contracts from an act specification. See the test for some examples.
