+++ act/act-programs/simple.act
@@ -1,0 +1,18 @@
+
+constructor of A
+interface constructor()
+
+creates
+
+  mapping (uint=>uint) x := []
+
+behaviour f of A
+interface f()
+
+storage
+
+  x[0] => 1
+
+returns
+
+  1
