+++ act/src/Act.hs
@@ -1,0 +1,5 @@
+module Act (module Act.Prelude, module Act.Execution, module Act.TH) where
+
+import Act.Execution
+import Act.Prelude
+import Act.TH
